Add files via upload
got the scheduler func working. 
created a generator to keep track of the last emp checked and start on that emp the next time the for loop starts. 
added doc strings.
created a job has no match boolean to check for jobs that would not fit anyone's schedule. hold them in a list and add them back to the heap at the end.